infra/gemini_desktop_commander/requirements.txt
google-generativeai>=0.5.0
pyautogui
pillow
screeninfo
pytesseract
python-dotenv
